Improved handling of multiple writes for same day
Bazefetcher output creates one file per day. Multiple subsequent writes
to the same file caused previously written content in this file to be
deleted.

This patch will allow multiple disjoint writes for a given day.

In the case of multiple overlapping writes, the user can specify if this
should result in an exception or an overwrite of the overlapping part.